linux-kernel: use lower Priority: values
no point in declaring any of these as Priority: required. the main
kernel package isn't even 'Required', because there may be other kernel
packages that are a better fit for a device. We'll let
spacemonkey-device and spacemonkey-current-kernel take care of that
stuff. and nothing of a higher priority depends on these packages, so
we're clear on that front.

mark the actual kernel packages as Priority: standard and the
headers/dbg packages as Priority: extra.

Change-Id: I4453ab10446da3ef7dd9907fa13ccf149bc36f53
Reviewed-on: https://gerrit.spacemonkey.com/5772
Reviewed-by: JT Olds <jt@spacemonkey.com>
Reviewed-by: Tom Stokes <tom@spacemonkey.com>
Tested-by: paul cannon <paul@spacemonkey.com>